Binded PrimaryButtonCommand to ViewModel
