+++ src/Files.App/Dialogs/CreateShortcutDialog.xaml
@@ -7,7 +7,7 @@
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	Title="{helpers:ResourceString Name=NewShortcutDialogTitle}"
 	DefaultButton="Primary"
-	PrimaryButtonClick="ContentDialog_PrimaryButtonClick"
+	PrimaryButtonCommand="{x:Bind ViewModel.PrimaryButtonCommand}"
 	IsPrimaryButtonEnabled="{x:Bind ViewModel.IsLocationValid, Mode=OneWay}"
 	PrimaryButtonText="{helpers:ResourceString Name=Create}"
 	RequestedTheme="{x:Bind helpers:ThemeHelper.RootTheme}"
