+++ src/Files.App/Dialogs/CreateShortcutDialog.xaml.cs
@@ -1,6 +1,4 @@
-using Files.App.Helpers;
 using Files.App.ViewModels.Dialogs;
-using Files.Backend.Extensions;
 using Files.Backend.ViewModels.Dialogs;
 using Files.Shared.Enums;
 using Microsoft.UI.Xaml.Controls;
@@ -12,8 +10,6 @@ namespace Files.App.Dialogs
 {
 	public sealed partial class CreateShortcutDialog : ContentDialog, IDialog<CreateShortcutDialogViewModel>
 	{
-		private bool _pathExists = false;
-
 		public CreateShortcutDialogViewModel ViewModel
 		{
 			get => (CreateShortcutDialogViewModel)DataContext;
@@ -23,39 +19,7 @@ namespace Files.App.Dialogs
 		public CreateShortcutDialog()
 		{
 			this.InitializeComponent();
-		}
-
-		private async void ContentDialog_PrimaryButtonClick(ContentDialog sender, ContentDialogButtonClickEventArgs args)
-		{
-			string? destinationName;
-			var extension = _pathExists ? ".lnk" : ".url";
-
-			if (_pathExists)
-			{
-				destinationName = Path.GetFileName(ViewModel.DestinationItemPath);
-				destinationName ??= Path.GetDirectoryName(ViewModel.DestinationItemPath);
-			}
-			else
-			{
-				var uri = new Uri(ViewModel.DestinationItemPath);
-				destinationName = uri.Host;
-			}
-
-			var shortcutName = string.Format("ShortcutCreateNewSuffix".ToLocalized(), destinationName);
-			var filePath = Path.Combine(
-				ViewModel.WorkingDirectory,
-				shortcutName + extension);
-
-			int fileNumber = 1;
-			while (Path.Exists(filePath))
-			{
-				filePath = Path.Combine(
-					ViewModel.WorkingDirectory,
-					shortcutName + $" ({++fileNumber})" + extension);
-			}
-
-			await FileOperationsHelpers.CreateOrUpdateLinkAsync(filePath, ViewModel.DestinationItemPath);
-		}
+		}	
 
 		public new async Task<DialogResult> ShowAsync() => (DialogResult)await base.ShowAsync();
 
@@ -69,8 +33,8 @@ namespace Files.App.Dialogs
 
 			try
 			{
-				_pathExists = Path.Exists(DestinationItemPath.Text) && DestinationItemPath.Text != Path.GetPathRoot(DestinationItemPath.Text);
-				if (_pathExists)
+				ViewModel.DestinationPathExists = Path.Exists(DestinationItemPath.Text) && DestinationItemPath.Text != Path.GetPathRoot(DestinationItemPath.Text);
+				if (ViewModel.DestinationPathExists)
 				{
 					ViewModel.IsLocationValid = true;
 				}
