+++ src/Files.App/ViewModels/Dialogs/CreateShortcutDialogViewModel.cs
@@ -1,6 +1,9 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using Files.App.Helpers;
+using Files.Backend.Extensions;
 using System;
+using System.IO;
 using System.Threading.Tasks;
 using System.Windows.Input;
 using Windows.Storage.Pickers;
@@ -11,6 +14,9 @@ namespace Files.App.ViewModels.Dialogs
 	{
 		// User's working directory
 		public readonly string WorkingDirectory;
+
+		// Tells whether destination path exists
+		public bool DestinationPathExists { get; set; } = false;
 
 		// Destination of the shortcut chosen by the user (can be a path or a URL)
 		private string _destinationItemPath;
@@ -31,12 +37,16 @@ namespace Files.App.ViewModels.Dialogs
 		// Command invoked when the user clicks the 'Browse' button
 		public ICommand SelectDestinationCommand { get; private set; }
 
+		// Command invoked when the user clicks primary button
+		public ICommand PrimaryButtonCommand { get; private set; }
+
 		public CreateShortcutDialogViewModel(string workingDirectory)
 		{
 			WorkingDirectory = workingDirectory;
 			_destinationItemPath = string.Empty;
 
 			SelectDestinationCommand = new AsyncRelayCommand(SelectDestination);
+			PrimaryButtonCommand = new AsyncRelayCommand(CreateShortcut);
 		}
 
 		private async Task SelectDestination()
@@ -53,6 +63,38 @@ namespace Files.App.ViewModels.Dialogs
 		{
 			WinRT.Interop.InitializeWithWindow.Initialize(obj, App.WindowHandle);
 			return obj;
+		}
+
+		private async Task CreateShortcut()
+		{
+			string? destinationName;
+			var extension = DestinationPathExists ? ".lnk" : ".url";
+
+			if (DestinationPathExists)
+			{
+				destinationName = Path.GetFileName(DestinationItemPath);
+				destinationName ??= Path.GetDirectoryName(DestinationItemPath);
+			}
+			else
+			{
+				var uri = new Uri(DestinationItemPath);
+				destinationName = uri.Host;
+			}
+
+			var shortcutName = string.Format("ShortcutCreateNewSuffix".ToLocalized(), destinationName);
+			var filePath = Path.Combine(
+				WorkingDirectory,
+				shortcutName + extension);
+
+			int fileNumber = 1;
+			while (Path.Exists(filePath))
+			{
+				filePath = Path.Combine(
+					WorkingDirectory,
+					shortcutName + $" ({++fileNumber})" + extension);
+			}
+
+			await FileOperationsHelpers.CreateOrUpdateLinkAsync(filePath, DestinationItemPath);
 		}
 	}
 }
